pyproject.toml
[project]
name = "nano-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai==1.12.1",
    "openai>=2.33.0",
    "python-dotenv==1.1.0",
]
[project.scripts]
nano-agent = "nano.main:main"
